Add revoked dialog
